TASK: Configure and render documents by name
This change adds a the configuration of the available spec documents to the setting and uses the document names as
basis for rendering via cli-command  and controller.

Th cli `./flow .openapi:document __name__` will render the same spec as `openapi/specs/__name__`.

Also the setting `schemaTargetFilePath` is removed. The cli will output the spec directly to the shell where it can be sent wherever it is needed.